requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.3            # via snakemake
attrs==19.2.0             # via jsonschema
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
configargparse==0.14.0    # via snakemake
datrie==0.8               # via snakemake
docutils==0.15.2          # via snakemake
gitdb2==2.0.6             # via gitpython
gitpython==3.0.3          # via snakemake
idna==2.8                 # via requests
jsonschema==3.0.2         # via snakemake
nltk==3.4.5
pyrsistent==0.15.4        # via jsonschema
python-dateutil==2.8.0
pyyaml==5.1.2
ratelimiter==1.2.0.post0  # via snakemake
requests==2.22.0          # via snakemake
six==1.12.0
smmap2==2.0.5             # via gitdb2
snakemake==5.13.0
urllib3==1.25.6           # via requests
wrapt==1.11.2             # via snakemake

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via jsonschema
